feat: mudando rota offers -> catalog
